Change source files structure for ZFS stats
